feat: add open close functionality
